Tidy up some ordered lists
Some of the lists had blank lines in them which, in Markdown, causes a
new list be rendered which isn't semantically correct, and numbering
will restart at 1, which is confusing.

There are some other minor reformatting and removal of <p> tags etc here
too as they appeared in/around the lists I was targetting.